Push added view models

+++ TrackerApp/Models/Hospital.cs
@@ -13,7 +13,7 @@ namespace EpidemicTracker.Data.Models
         public string Area { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public long Pincode { get; set; }
+        public int Pincode { get; set; }
         public string Country { get; set; }
         //public bool IsActive { get; set; }
         //public DateTime DateCreated { get; set; }
